+++ community/cypher/cypher-compiler-2.2/src/main/scala/org/neo4j/cypher/internal/compiler/v2_2/planner/logical/expandTableSolver.scala
@@ -19,26 +19,52 @@
  */
 package org.neo4j.cypher.internal.compiler.v2_2.planner.logical
 
+import org.neo4j.cypher.internal.compiler.v2_2.ast.{Identifier, FilterScope, AllIterablePredicate}
 import org.neo4j.cypher.internal.compiler.v2_2.planner.QueryGraph
-import org.neo4j.cypher.internal.compiler.v2_2.planner.logical.plans.LogicalPlan
+import org.neo4j.cypher.internal.compiler.v2_2.planner.logical.plans._
+import org.neo4j.cypher.internal.compiler.v2_2.planner.logical.steps.LogicalPlanProducer._
 
 object expandTableSolver extends ExhaustiveTableSolver {
 
-  import org.neo4j.cypher.internal.compiler.v2_2.planner.logical.ExhaustiveQueryGraphSolver.planSinglePatternSide
-
   override def apply(qg: QueryGraph, goal: Set[Solvable], table: (Set[Solvable]) => Option[LogicalPlan]): Iterator[LogicalPlan] = {
     val result =
-      for(
-        solvable <- goal.iterator;
-        pattern <- solvable.solvedRelationship;
-        solved = goal - solvable;
-        plan <- table(solved) // if !plan.solved.graph.patternRelationships(pattern)
+      for (solvable <- goal.iterator;
+           pattern <- solvable.solvedRelationship;
+           solved = goal - solvable;
+           plan <- table(solved)
       ) yield {
-          Iterator(
-            planSinglePatternSide(qg, pattern, plan, pattern.left),
-            planSinglePatternSide(qg, pattern, plan, pattern.right)
-          ).flatten
+        Iterator(
+          planSinglePatternSide(qg, pattern, plan, pattern.left),
+          planSinglePatternSide(qg, pattern, plan, pattern.right)
+        ).flatten
       }
       result.flatten
   }
+
+  def planSinglePatternSide(qg: QueryGraph, patternRel: PatternRelationship, plan: LogicalPlan, nodeId: IdName): Option[LogicalPlan] = {
+    val availableSymbols = plan.availableSymbols
+    if (availableSymbols(nodeId)) {
+      val dir = patternRel.directionRelativeTo(nodeId)
+      val otherSide = patternRel.otherSide(nodeId)
+      val overlapping = availableSymbols.contains(otherSide)
+      val mode = if (overlapping) ExpandInto else ExpandAll
+
+      patternRel.length match {
+        case SimplePatternLength =>
+          Some(planSimpleExpand(plan, nodeId, dir, otherSide, patternRel, mode))
+
+        case length: VarPatternLength =>
+          val availablePredicates = qg.selections.predicatesGiven(availableSymbols + patternRel.name)
+          val (predicates, allPredicates) = availablePredicates.collect {
+            case all@AllIterablePredicate(FilterScope(identifier, Some(innerPredicate)), relId@Identifier(patternRel.name.name))
+              if identifier == relId || !innerPredicate.dependencies(relId) =>
+              (identifier, innerPredicate) -> all
+          }.unzip
+          Some(planVarExpand(plan, nodeId, dir, otherSide, patternRel, predicates, allPredicates, mode))
+      }
+    } else {
+      None
+    }
+  }
+
 }
